Allow to restore LP solution after diving without resolving the LP.
If new parameter "lp/resolverestore" is set to FALSE then we buffer
the LP solution values in the lp data, in each column and row and
restore those at end of diving.  Otherwise, we resolve the LP after
loading the lp state.

If the new parameter "lp/freesolvalbuffers" is set to TRUE (FALSE by
default), we free the memory for storing the LP solution values at the
end of each dive to save memory.