eliminate min coef in cut if coef range too large
If range of coefficients is too large, then this may be due to a
very small minimal coefficient (e.g., 2e-9). We then try to increase
the minimal coefficient by eliminating the coefficient, thereby
weakening the cut.
This may work well if the small coefficient is due to an almost
fixed variable (was the case for nous2).