add possibility to link a SCIP_SOL to the current NLP solution
The SCIP_VAR now has an entry to store the value of the variable in
the current NLP (analog to solutions from an arbitrary relaxation).
This entry is updated by the NLP and can be requested by SCIPvarGetNLPSol,
which replaces SCIPnlpGetVarSolVal.

Additionally (and to make implementation of the sol linking easier),
we drop the idea of using a different objective than the SCIP objective
in the NLP. During NLP diving, this is still possible.