move analysis of probing deductions into separate function
Added public function SCIPanalyzeDeductionsProbing to perform
analysis of bound tightenings found by probing some variable on
some bounds. The probing variable does not need to be binary here,
even though implications are only added if it is binary.

The case where aggregations are added is slightly generalized, i.e.,
the involved variables do not have to be binary anymore.