Updated project and shift to more efficiently prove node infeasibility
Project and shift can now prove node infeasibility by correcting a dual ray
to be exactly feasible.  This is activated by setting
misc/psinfeasray = TRUE.  This improves the speed when many nodes are
infeasible and no primal bound exists.  Also, a bug was fixed where the
incorrect lambda shifting value could be computed when using the setting
constraints/exactlp/psuseintpoint = FALSE - this would cause a point to not
be corrected to be feasible and would cause an assert in debug mode.