add yet another method to compute relative interior of current LP
Renamed SCIPcomputeLPRelIntPoint() to SCIPcomputeLPRelIntPointOneNorm()
and added function SCIPcomputeLPRelIntPointSupNorm() to compute
relative interior of the current LP w.r.t. supremum-norm, i.e.,
maximize the supremum of the row and column bounds slacks.
The new method does not use homogenization yet and thus may not work
if the polyhedra is unbounded.

Added parameter to closecuts separator to select type of norm for
computing relative interior point.