keep varbound constraints as initial in reformulation
It seems that when reformulating a product with binary variables,
keeping the varbound constraints as initial and only making the linear
constraints non-initial is advantageous in some cases.

We loose a bit on iair04, imisc07, inug06-3rd, iran8x32, iswath2, but
also improve on iair05, ibc1, ibienst1, icap6000, iqap10, iqiu, iran13x13.
Further, the root dual bound for lop97icx becomes almost as good as if
one makes also the linear constraints initial.