add writing of indicator constraints to gams files
Since indicators are not allowed by the GAMS language (except for some
hack in the GAMS/CPLEX link), two reformulations are available:
Either by using a big-M reformulation (which requires a big M, obviously),
or via special ordered sets of type 1. The latter introduces 2 artificial
variables an equations for each indicator constraint due to GAMS
restrictive way on formulating SOS, but at least some solvers should be
able to presolve these additions away.
Further, note that SOS is by far not understand by all solvers that are
available in GAMS.

When printing the transformed problem, add the objective offset to the
equation that defines the objective function.