revised handling of fixing of upper right triangle
The fixing of the upper right triangle has been moved
(copied)  to exitpre (newly implemented). This guarantees
that it is now always fixed and allowed the removal of some
code. In particular, computeSCTable()
could be removed (computeSCTableFixTriangle() is renamed
to computeSCTable()). SCIPcheck() has been modified
to check the upper right triangle.