Assert finiteness for aggregated vars in SCIPvarGetLPSol_rec(), ...
..., SCIPvarGetPseudoSol_rec(), SCIPvarGetRootSol(),
SCIPvarGetLbAtIndex(), and SCIPvarGetUbAtIndex().

A correct implementation would need to check the value of
var->data.aggregate.var for infinity and return the corresponding
infinity value instead of performing an arithmetical transformation
(compare method SCIPvarGetLbLP()); however, we do not want to
introduce a SCIP or SCIP_SET pointer to this method, since it is (or
is called by) a public interface method; instead, we only assert that
values are finite w.r.t. SCIP_DEFAULT_INFINITY, which seems to be true
in our regression tests; note that this may yield false positives and
negatives if the parameter <numerics/infinity> is modified by the
user.