Always clear LP state after probing mode if LP was initially unsolved.
If during probing mode the LP relaxation is solved from scratch then
we clear the resulting LP state, since it might be a bad starting
basis for the next solve of the LP relaxation. Controlled by new
parameter <lp/clearinitialprobinglp>, TRUE by default.

This reverts commit 7e65a138518354d405b1596488de8d3ab2f7a544 where we
introduced the parameter <heuristics/shiftandpropagate/clearstate> to
achieve this only for this heuristic - we now perform the clear in
tree.c.