remove solution of imod011 for MINLP.solu
A point with objective -8.8239 had been reported by cplex, but a closer
look shows small infeasibilities which do not seem to be fixable while
keeping the objective value.
That is, the solution reported by cplex has 1e-8 for several binary
variables, and 0 for all other ones. Fixing all binary variables to 0
and solving the remaining convex QCP by CONOPT (starting from the
solution values for the continuous variables) yields the all-zero
point as optimal solution.
Thus, we replace the =opt= entry in the solu file for imod011 by
=best= 0.0