fix todo "this has to be corrected for lb = -infinity"
When loading an old basis state in SCIPlpiSetState() new columns
might have been added in the meantime. Previously these columns
were always set to be at their lower bounds. If the new variables do
not have finite lower bounds, now either the upper bounds are
taken or the variables are free/super basic. In MOSEK and Xpress
this part was missing altogether, which is now implemented.