Create bound changes instead of linear constraints
The nonlinear constraint handlers generate lineare constraints
if all "nonlinearities are fixed". We now check whether this would
result in a linear constraint with one variable and perform a bound
change in this case.

As a side effect this avoids bug 362, which resulted from a linear
constraint with one variable that was enforced as a bound change
and because of numerical issues still lead to infeasibility, since
the bound change was not accepted. See the explanation there.