added reading of negated and fixed variables; resort output
Fixed and negated variables can now in principle be read from
CIP-files. However, we need to make sure that all variables
used in aggregations/negations are defined before.
As a heuristic we first write the fixed variables, then negated
and aggregated variables, and then multi-aggregated variables.
This has to be fixed.