[#1093] It may happen that setting dual norms fails.
Note that this is usually not critical. In the case at hand, the issue
was that the model was still solved to optimality. In this case, SCIP
does not call SCIPlpiSetState() again, and this means that there is no
external warm start information in Gurobi. And since dual norms can
only be loaded if an external warm start is available, this results in
a Gurobi error.