some fixes in cmake files
do not built static sub libraries because they cannot be linked into
other static libraries with cmake.
Cmake does not allow this because somme compilers do not support this
and therefore it is not possible in a portable way.
This fixes errors when building a static scipopt library.
Additionally the libraries that where used for zimpl and soplex
are not packaged with the scip library when building in static mode.
Therefore the scip config file now quietly searches for the ZIMPL and SoPlex
libraries that where used to build the scip library.
Now the scipopt library can be used from external cmake projects just by using
find_package(SCIP) and it works for both, static and shared scipopt libraries.