changed assert which was could be thrown by arithmetic precision
If value is just outside of the epsilon tolerance of an integer by 10^-16
or less, it may happen due to arithmetic errors, that the number is still
reported as integral, while the absolute value is not. Because we add
feastol before rounding down later anyways, it suffices to check the
absolute value for feasible integrality.