Revert "relax equality test in prop_dualfix"
This reverts commit 79e5d2311f64c029bae623c91a56b09d6ad9a21d.

It is true that SCIPfixVar will only apply boundchanges that are more
than feastol-better than the previous one, but only if in solving or
probing. If in presolve, then SCIPfixVar would always fix the variable,
i.e., change it from active to fixed.

Note, further, that 90232102 (committed after 79e5d2311) ensures that the
assert(fixed) that was there before does not fail if solving and the bound
change is below feastol.
So 90232102 is kind of an alternative fix to 79e5d2311.