fix cons_nonlinear consssorted flag for variables
src/scip/cons_nonlinear.c (catchVarEvent): Fix inverted logic when
setting ownerdata->consssorted for a variable. This resulted in the
array incorrectly being marked as sorted when it was actually
reverse-sorted, and as a result, the lookup with SCIPsortedvecFindPtr in
dropVarEvent failing with:
[cons_nonlinear.c:1250] ERROR: Constraint <...> not in constraint array
of expression for variable <...>
Thankfully, the fix is trivial.

Fixes scipopt/scip#147.