WIP/RFC: Do not call SCIPincludeObjExprhdlr in exprCopyhdlrObj
src/objscip/objexprhdlr.cpp (exprCopyhdlrObj): Comment out the call to
SCIPincludeObjExprhdlr because this is called from the Java code.

See https://github.com/scipopt/scip/pull/149#issuecomment-2986486164 for
details.

TODO: This is a quick hack to fix the double free. Find a solution that
is acceptable upstream and gets applied consistently also to constraint
handlers etc.