fix cons_nonlinear SCIPaddExprsViolScoreNonlinear crash
src/scip/cons_nonlinear.c (SCIPaddExprsViolScoreNonlinear): Do not call
addExprsViolScore with nexprs=nvars if nvars==0. addExprsViolScore
assumes that nexprs>0, and always tries to access the first expression,
which will crash if the array of expressions is empty (or if assertions
are enabled, it will fail the nexprs > 0 assertion instead). So if the
expressions had no variables (nvars==0), proceed the same way as if we
were passed no expressions to begin with (nexprs==0): set *success to
FALSE and return SCIP_OKAY, avoiding the call to addExprsViolScore.